requirements.txt
# Core ML and NLP Libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Vector Database and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.4

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Web Framework
streamlit==1.29.0
gradio==4.7.1

# Explainability and Visualization
shap==0.44.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
wordcloud==1.9.3

# Utilities
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# German Language Support
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0.tar.gz

# Additional NLP Tools
nltk==3.8.1
textblob==0.17.1

# Database (optional)
sqlite3

# Logging
loguru==0.7.2

# Async Support
aiofiles==23.2.1
asyncio

# Memory Management
psutil==5.9.6

# Security
cryptography==41.0.8 